Remove one or more score specifications

Delete one or more rows from a scores table, where each row represents a score specification.

Usage

delete_scores(scores, rows)

Arguments

scores

A data frame of score specifications. See scores_init.

rows

A numeric vector of rows to remove from the table.

Value

scores with some rows removed.

See also

Examples

scores <- create_score(
  scores_init, score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)

delete_scores(scores, 1)
#> # A tibble: 0 × 12
#> # … with 12 variables: score_type <chr>, colname <chr>, score_name <glue>,
#> #   weight <dbl>, lb <dbl>, ub <dbl>, centre <dbl>, inverse <lgl>,
#> #   exponential <lgl>, logarithmic <lgl>, magnitude <dbl>, custom_args <list>