Model the residuals of stock price predictions.

Get the daily or monthly LightGBM models, which are trained to predict the residuals of the Prophet model.

Usage

get_lightgbm_model(x)

Arguments

x

Which model to get: either "daily" or "monthly".

Value

A fitted workflows::workflow() object.

Details

This model is designed to perform iterative forecasting. This means that it can predict residuals for a certain horizon, then assumes these predictions as truth and uses lagged values as predictors, and makes new predictions, and so on.

The process to get the models is complex because readRDS() does not work for LightGBM models.

See also

Examples

get_lightgbm_model("daily")
#> $pre
#> $actions
#> $actions$recipe
#> $recipe
#> $var_info
#> # A tibble: 86 × 4
#>    variable        type    role      source  
#>    <chr>           <chr>   <chr>     <chr>   
#>  1 ticker          nominal predictor original
#>  2 ds              date    predictor original
#>  3 residuals_lag8  numeric predictor original
#>  4 residuals_lag9  numeric predictor original
#>  5 residuals_lag10 numeric predictor original
#>  6 residuals_lag11 numeric predictor original
#>  7 residuals_lag12 numeric predictor original
#>  8 residuals_lag13 numeric predictor original
#>  9 residuals_lag14 numeric predictor original
#> 10 residuals_lag15 numeric predictor original
#> # … with 76 more rows
#> 
#> $term_info
#> # A tibble: 86 × 4
#>    variable        type    role      source  
#>    <chr>           <chr>   <chr>     <chr>   
#>  1 ticker          nominal predictor original
#>  2 ds              date    predictor original
#>  3 residuals_lag8  numeric predictor original
#>  4 residuals_lag9  numeric predictor original
#>  5 residuals_lag10 numeric predictor original
#>  6 residuals_lag11 numeric predictor original
#>  7 residuals_lag12 numeric predictor original
#>  8 residuals_lag13 numeric predictor original
#>  9 residuals_lag14 numeric predictor original
#> 10 residuals_lag15 numeric predictor original
#> # … with 76 more rows
#> 
#> $steps
#> $steps[[1]]
#> $terms
#> $terms[[1]]
#> <quosure>
#> expr: ^ds
#> env:  package:base
#> 
#> 
#> $role
#> [1] "predictor"
#> 
#> $trained
#> [1] FALSE
#> 
#> $features
#> [1] "month"   "dow"     "quarter" "week"    "decimal"
#> 
#> $abbr
#> [1] TRUE
#> 
#> $label
#> [1] TRUE
#> 
#> $ordinal
#> [1] FALSE
#> 
#> $locale
#> [1] "en_US.UTF-8"
#> 
#> $columns
#> NULL
#> 
#> $keep_original_cols
#> [1] TRUE
#> 
#> $skip
#> [1] FALSE
#> 
#> $id
#> [1] "date_seuTv"
#> 
#> attr(,"class")
#> [1] "step_date" "step"     
#> 
#> $steps[[2]]
#> $terms
#> $terms[[1]]
#> <quosure>
#> expr: ^ds
#> env:  package:base
#> 
#> 
#> $role
#> [1] "predictor"
#> 
#> $trained
#> [1] FALSE
#> 
#> $frequency
#> [1] 1 2 3 4 5
#> 
#> $cycle_size
#> [1] 30
#> 
#> $starting_val
#> [1] NA
#> 
#> $keep_original_cols
#> [1] TRUE
#> 
#> $columns
#> NULL
#> 
#> $skip
#> [1] FALSE
#> 
#> $id
#> [1] "harmonic_27pMR"
#> 
#> attr(,"class")
#> [1] "step_harmonic" "step"         
#> 
#> $steps[[3]]
#> $terms
#> $terms[[1]]
#> <quosure>
#> expr: ^ds
#> env:  package:base
#> 
#> 
#> $role
#> [1] NA
#> 
#> $trained
#> [1] FALSE
#> 
#> $removals
#> NULL
#> 
#> $skip
#> [1] FALSE
#> 
#> $id
#> [1] "rm_RQuNv"
#> 
#> attr(,"class")
#> [1] "step_rm" "step"   
#> 
#> 
#> $template
#> # A tibble: 0 × 86
#> # … with 86 variables: ticker <fct>, ds <date>, residuals_lag8 <dbl>,
#> #   residuals_lag9 <dbl>, residuals_lag10 <dbl>, residuals_lag11 <dbl>,
#> #   residuals_lag12 <dbl>, residuals_lag13 <dbl>, residuals_lag14 <dbl>,
#> #   residuals_lag15 <dbl>, residuals_lag16 <dbl>, residuals_lag17 <dbl>,
#> #   residuals_lag18 <dbl>, residuals_lag19 <dbl>, residuals_lag20 <dbl>,
#> #   residuals_lag21 <dbl>, residuals_lag22 <dbl>, residuals_lag23 <dbl>,
#> #   residuals_lag24 <dbl>, residuals_lag25 <dbl>, residuals_lag26 <dbl>, …
#> 
#> $levels
#> NULL
#> 
#> $retained
#> [1] NA
#> 
#> $requirements
#> $requirements$bake
#> named logical(0)
#> 
#> 
#> attr(,"class")
#> [1] "butchered_recipe" "recipe"          
#> 
#> $blueprint
#> $intercept
#> [1] FALSE
#> 
#> $allow_novel_levels
#> [1] FALSE
#> 
#> $composition
#> [1] "tibble"
#> 
#> $ptypes
#> NULL
#> 
#> $fresh
#> [1] TRUE
#> 
#> $recipe
#> NULL
#> 
#> $extra_role_ptypes
#> NULL
#> 
#> attr(,"class")
#> [1] "default_recipe_blueprint" "recipe_blueprint"        
#> [3] "hardhat_blueprint"       
#> 
#> attr(,"class")
#> [1] "action_recipe" "action_pre"    "action"       
#> 
#> 
#> $mold
#> $mold$blueprint
#> $intercept
#> [1] FALSE
#> 
#> $allow_novel_levels
#> [1] FALSE
#> 
#> $composition
#> [1] "tibble"
#> 
#> $ptypes
#> $ptypes$predictors
#> # A tibble: 0 × 85
#> # … with 85 variables: ticker <fct>, ds <date>, residuals_lag8 <dbl>,
#> #   residuals_lag9 <dbl>, residuals_lag10 <dbl>, residuals_lag11 <dbl>,
#> #   residuals_lag12 <dbl>, residuals_lag13 <dbl>, residuals_lag14 <dbl>,
#> #   residuals_lag15 <dbl>, residuals_lag16 <dbl>, residuals_lag17 <dbl>,
#> #   residuals_lag18 <dbl>, residuals_lag19 <dbl>, residuals_lag20 <dbl>,
#> #   residuals_lag21 <dbl>, residuals_lag22 <dbl>, residuals_lag23 <dbl>,
#> #   residuals_lag24 <dbl>, residuals_lag25 <dbl>, residuals_lag26 <dbl>, …
#> 
#> $ptypes$outcomes
#> # A tibble: 0 × 1
#> # … with 1 variable: residuals <dbl>
#> 
#> 
#> $fresh
#> [1] TRUE
#> 
#> $recipe
#> $var_info
#> # A tibble: 86 × 4
#>    variable        type    role      source  
#>    <chr>           <chr>   <chr>     <chr>   
#>  1 ticker          nominal predictor original
#>  2 ds              date    predictor original
#>  3 residuals_lag8  numeric predictor original
#>  4 residuals_lag9  numeric predictor original
#>  5 residuals_lag10 numeric predictor original
#>  6 residuals_lag11 numeric predictor original
#>  7 residuals_lag12 numeric predictor original
#>  8 residuals_lag13 numeric predictor original
#>  9 residuals_lag14 numeric predictor original
#> 10 residuals_lag15 numeric predictor original
#> # … with 76 more rows
#> 
#> $term_info
#> # A tibble: 100 × 4
#>    variable        type    role      source  
#>    <chr>           <chr>   <chr>     <chr>   
#>  1 ticker          nominal predictor original
#>  2 residuals_lag8  numeric predictor original
#>  3 residuals_lag9  numeric predictor original
#>  4 residuals_lag10 numeric predictor original
#>  5 residuals_lag11 numeric predictor original
#>  6 residuals_lag12 numeric predictor original
#>  7 residuals_lag13 numeric predictor original
#>  8 residuals_lag14 numeric predictor original
#>  9 residuals_lag15 numeric predictor original
#> 10 residuals_lag16 numeric predictor original
#> # … with 90 more rows
#> 
#> $steps
#> $steps[[1]]
#> $terms
#> $terms[[1]]
#> <quosure>
#> expr: ^ds
#> env:  package:base
#> 
#> 
#> $role
#> [1] "predictor"
#> 
#> $trained
#> [1] TRUE
#> 
#> $features
#> [1] "month"   "dow"     "quarter" "week"    "decimal"
#> 
#> $abbr
#> [1] TRUE
#> 
#> $label
#> [1] TRUE
#> 
#> $ordinal
#> [1] FALSE
#> 
#> $locale
#> [1] "en_US.UTF-8"
#> 
#> $columns
#>   ds 
#> "ds" 
#> 
#> $keep_original_cols
#> [1] TRUE
#> 
#> $skip
#> [1] FALSE
#> 
#> $id
#> [1] "date_seuTv"
#> 
#> attr(,"class")
#> [1] "step_date" "step"     
#> 
#> $steps[[2]]
#> $terms
#> $terms[[1]]
#> <quosure>
#> expr: ^ds
#> env:  package:base
#> 
#> 
#> $role
#> [1] "predictor"
#> 
#> $trained
#> [1] TRUE
#> 
#> $frequency
#> 1 2 3 4 5 
#> 1 2 3 4 5 
#> 
#> $cycle_size
#> ds 
#> 30 
#> 
#> $starting_val
#> ds 
#>  0 
#> 
#> $keep_original_cols
#> [1] TRUE
#> 
#> $columns
#>   ds 
#> "ds" 
#> 
#> $skip
#> [1] FALSE
#> 
#> $id
#> [1] "harmonic_27pMR"
#> 
#> attr(,"class")
#> [1] "step_harmonic" "step"         
#> 
#> $steps[[3]]
#> $terms
#> $terms[[1]]
#> <quosure>
#> expr: ^ds
#> env:  package:base
#> 
#> 
#> $role
#> [1] NA
#> 
#> $trained
#> [1] TRUE
#> 
#> $removals
#>   ds 
#> "ds" 
#> 
#> $skip
#> [1] FALSE
#> 
#> $id
#> [1] "rm_RQuNv"
#> 
#> attr(,"class")
#> [1] "step_rm" "step"   
#> 
#> 
#> $levels
#> $levels$ticker
#> $levels$ticker$values
#>   [1] "A"     "AAL"   "AAP"   "AAPL"  "ABBV"  "ABC"   "ABMD"  "ABT"   "ACN"  
#>  [10] "ADBE"  "ADI"   "ADM"   "ADP"   "ADSK"  "AEE"   "AEP"   "AES"   "AFL"  
#>  [19] "AIG"   "AIZ"   "AJG"   "AKAM"  "ALB"   "ALGN"  "ALK"   "ALL"   "ALLE" 
#>  [28] "AMAT"  "AMCR"  "AMD"   "AME"   "AMGN"  "AMP"   "AMT"   "AMZN"  "ANET" 
#>  [37] "ANSS"  "AON"   "AOS"   "APA"   "APD"   "APH"   "APTV"  "ARE"   "ATO"  
#>  [46] "ATVI"  "AVB"   "AVGO"  "AVY"   "AWK"   "AXP"   "AZO"   "BA"    "BAC"  
#>  [55] "BALL"  "BAX"   "BBWI"  "BBY"   "BDX"   "BIIB"  "BIO"   "BK"    "BKNG" 
#>  [64] "BKR"   "BLK"   "BMY"   "BR"    "BRO"   "BSX"   "BWA"   "BXP"   "CAH"  
#>  [73] "CARR"  "CAT"   "CBOE"  "CBRE"  "CCL"   "CDAY"  "CDNS"  "CDW"   "CE"   
#>  [82] "CHRW"  "CNC"   "CNP"   "COF"   "CPB"   "CPT"   "CTLT"  "CZR"   "GOOG" 
#>  [91] "GOOGL" "KMX"   "LNT"   "MMM"   "MO"    "T"     "TECH"  "WRB"   "C"    
#> [100] "CAG"   "CB"    "CCI"   "CEG"   "CF"    "CFG"   "CHD"   "CHTR"  "CI"   
#> [109] "CINF"  "CL"    "CLX"   "CMA"   "CMCSA" "CME"   "CMG"   "CMI"   "CMS"  
#> [118] "COO"   "COP"   "COST"  "CPRT"  "CRL"   "CSCO"  "CSX"   "CTAS"  "CTRA" 
#> [127] "CTSH"  "CTVA"  "CTXS"  "CVS"   "CVX"   "D"     "DAL"   "DD"    "DE"   
#> [136] "DFS"   "DG"    "DHI"   "DHR"   "DIS"   "DISH"  "DLR"   "DLTR"  "DOV"  
#> [145] "DOW"   "DPZ"   "DRE"   "DRI"   "DTE"   "DUK"   "DVA"   "DVN"   "DXC"  
#> [154] "DXCM"  "EA"    "EBAY"  "ECL"   "ED"    "EFX"   "EIX"   "EL"    "ELV"  
#> [163] "EMN"   "EMR"   "ENPH"  "EOG"   "EPAM"  "EQIX"  "EQR"   "ES"    "ESS"  
#> [172] "ETN"   "ETR"   "ETSY"  "EVRG"  "EW"    "EXC"   "EXPD"  "EXPE"  "EXR"  
#> [181] "FANG"  "FAST"  "FDS"   "FDX"   "FE"    "FFIV"  "FIS"   "FITB"  "FRC"  
#> [190] "FRT"   "GLW"   "KO"    "LLY"   "RE"    "SCHW"  "STZ"   "XOM"   "XRAY" 
#> [199] "BEN"   "F"     "FBHS"  "FCX"   "FISV"  "FLT"   "FMC"   "FOX"   "FOXA" 
#> [208] "FTNT"  "FTV"   "GD"    "GE"    "GILD"  "GIS"   "GL"    "GM"    "GNRC" 
#> [217] "GPC"   "GPN"   "GRMN"  "GS"    "HAL"   "HAS"   "HBAN"  "HCA"   "HD"   
#> [226] "HES"   "HIG"   "HII"   "HLT"   "HOLX"  "HON"   "HPE"   "HPQ"   "HRL"  
#> [235] "HSIC"  "HST"   "HSY"   "HUM"   "HWM"   "IBM"   "ICE"   "IDXX"  "IEX"  
#> [244] "IFF"   "ILMN"  "INCY"  "INTC"  "INTU"  "IP"    "IPG"   "IQV"   "IR"   
#> [253] "IRM"   "ISRG"  "IT"    "ITW"   "IVZ"   "J"     "JBHT"  "JCI"   "JKHY" 
#> [262] "JNJ"   "JNPR"  "JPM"   "K"     "KDP"   "KEY"   "KEYS"  "KHC"   "KIM"  
#> [271] "KLAC"  "KMB"   "KMI"   "KR"    "L"     "LDOS"  "LEN"   "LH"    "LHX"  
#> [280] "LIN"   "LKQ"   "LMT"   "LNC"   "LOW"   "LRCX"  "LUMN"  "LVS"   "LW"   
#> [289] "LYB"   "LYV"   "MAR"   "MKTX"  "MLM"   "MMC"   "MPC"   "MRO"   "MTB"  
#> [298] "PEAK"  "DGX"   "MA"    "MAA"   "MAS"   "MCD"   "MCHP"  "MCK"   "MCO"  
#> [307] "MDLZ"  "MDT"   "MET"   "META"  "MGM"   "MHK"   "MKC"   "MNST"  "MOH"  
#> [316] "MOS"   "MPWR"  "MRK"   "MRNA"  "MS"    "MSCI"  "MSFT"  "MSI"   "MTCH" 
#> [325] "MTD"   "MU"    "NCLH"  "NDAQ"  "NDSN"  "NEE"   "NEM"   "NFLX"  "NI"   
#> [334] "NKE"   "NLOK"  "NLSN"  "NOC"   "NRG"   "NSC"   "NTAP"  "NTRS"  "NUE"  
#> [343] "NVDA"  "NVR"   "NWL"   "NWS"   "NWSA"  "NXPI"  "O"     "ODFL"  "OGN"  
#> [352] "OKE"   "OMC"   "ON"    "ORCL"  "ORLY"  "OTIS"  "OXY"   "PARA"  "PAYC" 
#> [361] "PAYX"  "PCAR"  "PEG"   "PENN"  "PEP"   "PFE"   "PFG"   "PG"    "PGR"  
#> [370] "PH"    "PHM"   "PKG"   "PKI"   "PLD"   "PM"    "PNC"   "PNR"   "PNW"  
#> [379] "POOL"  "PPG"   "PPL"   "PRU"   "PSA"   "PSX"   "PTC"   "PVH"   "PWR"  
#> [388] "PXD"   "PYPL"  "QCOM"  "QRVO"  "REG"   "REGN"  "RF"    "RJF"   "RL"   
#> [397] "RTX"   "TAP"   "CRM"   "GWW"   "LUV"   "NOW"   "RCL"   "RHI"   "RMD"  
#> [406] "ROK"   "ROL"   "ROP"   "ROST"  "RSG"   "SBAC"  "SBNY"  "SBUX"  "SEDG" 
#> [415] "SEE"   "SHW"   "SIVB"  "SJM"   "SLB"   "SNA"   "SNPS"  "SO"    "SPG"  
#> [424] "SPGI"  "SRE"   "STE"   "STT"   "STX"   "SWK"   "SWKS"  "SYF"   "SYK"  
#> [433] "SYY"   "TDG"   "TDY"   "TEL"   "TER"   "TFC"   "TFX"   "TGT"   "TJX"  
#> [442] "TMO"   "TMUS"  "TPR"   "TRMB"  "TROW"  "TRV"   "TSCO"  "TSLA"  "TSN"  
#> [451] "TT"    "TTWO"  "TWTR"  "TXN"   "TXT"   "TYL"   "UAL"   "UDR"   "UHS"  
#> [460] "ULTA"  "UNH"   "UNP"   "UPS"   "URI"   "USB"   "V"     "VFC"   "VICI" 
#> [469] "VLO"   "VMC"   "VNO"   "VRSK"  "VRSN"  "VRTX"  "VTR"   "VTRS"  "VZ"   
#> [478] "WAB"   "WAT"   "WBA"   "WBD"   "WDC"   "WEC"   "WELL"  "WFC"   "WHR"  
#> [487] "WM"    "WMB"   "WMT"   "WRK"   "WST"   "WTW"   "WY"    "WYNN"  "XEL"  
#> [496] "XYL"   "YUM"   "ZBH"   "ZBRA"  "ZION"  "ZTS"  
#> 
#> $levels$ticker$ordered
#> [1] FALSE
#> 
#> $levels$ticker$factor
#> [1] TRUE
#> 
#> 
#> $levels$residuals_lag8
#> $levels$residuals_lag8$values
#> [1] NA
#> 
#> $levels$residuals_lag8$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag9
#> $levels$residuals_lag9$values
#> [1] NA
#> 
#> $levels$residuals_lag9$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag10
#> $levels$residuals_lag10$values
#> [1] NA
#> 
#> $levels$residuals_lag10$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag11
#> $levels$residuals_lag11$values
#> [1] NA
#> 
#> $levels$residuals_lag11$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag12
#> $levels$residuals_lag12$values
#> [1] NA
#> 
#> $levels$residuals_lag12$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag13
#> $levels$residuals_lag13$values
#> [1] NA
#> 
#> $levels$residuals_lag13$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag14
#> $levels$residuals_lag14$values
#> [1] NA
#> 
#> $levels$residuals_lag14$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag15
#> $levels$residuals_lag15$values
#> [1] NA
#> 
#> $levels$residuals_lag15$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag16
#> $levels$residuals_lag16$values
#> [1] NA
#> 
#> $levels$residuals_lag16$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag17
#> $levels$residuals_lag17$values
#> [1] NA
#> 
#> $levels$residuals_lag17$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag18
#> $levels$residuals_lag18$values
#> [1] NA
#> 
#> $levels$residuals_lag18$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag19
#> $levels$residuals_lag19$values
#> [1] NA
#> 
#> $levels$residuals_lag19$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag20
#> $levels$residuals_lag20$values
#> [1] NA
#> 
#> $levels$residuals_lag20$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag21
#> $levels$residuals_lag21$values
#> [1] NA
#> 
#> $levels$residuals_lag21$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag22
#> $levels$residuals_lag22$values
#> [1] NA
#> 
#> $levels$residuals_lag22$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag23
#> $levels$residuals_lag23$values
#> [1] NA
#> 
#> $levels$residuals_lag23$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag24
#> $levels$residuals_lag24$values
#> [1] NA
#> 
#> $levels$residuals_lag24$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag25
#> $levels$residuals_lag25$values
#> [1] NA
#> 
#> $levels$residuals_lag25$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag26
#> $levels$residuals_lag26$values
#> [1] NA
#> 
#> $levels$residuals_lag26$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag27
#> $levels$residuals_lag27$values
#> [1] NA
#> 
#> $levels$residuals_lag27$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag28
#> $levels$residuals_lag28$values
#> [1] NA
#> 
#> $levels$residuals_lag28$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag29
#> $levels$residuals_lag29$values
#> [1] NA
#> 
#> $levels$residuals_lag29$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag30
#> $levels$residuals_lag30$values
#> [1] NA
#> 
#> $levels$residuals_lag30$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag31
#> $levels$residuals_lag31$values
#> [1] NA
#> 
#> $levels$residuals_lag31$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag32
#> $levels$residuals_lag32$values
#> [1] NA
#> 
#> $levels$residuals_lag32$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag33
#> $levels$residuals_lag33$values
#> [1] NA
#> 
#> $levels$residuals_lag33$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag34
#> $levels$residuals_lag34$values
#> [1] NA
#> 
#> $levels$residuals_lag34$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag35
#> $levels$residuals_lag35$values
#> [1] NA
#> 
#> $levels$residuals_lag35$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag36
#> $levels$residuals_lag36$values
#> [1] NA
#> 
#> $levels$residuals_lag36$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag37
#> $levels$residuals_lag37$values
#> [1] NA
#> 
#> $levels$residuals_lag37$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag38
#> $levels$residuals_lag38$values
#> [1] NA
#> 
#> $levels$residuals_lag38$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag39
#> $levels$residuals_lag39$values
#> [1] NA
#> 
#> $levels$residuals_lag39$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag40
#> $levels$residuals_lag40$values
#> [1] NA
#> 
#> $levels$residuals_lag40$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag41
#> $levels$residuals_lag41$values
#> [1] NA
#> 
#> $levels$residuals_lag41$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag42
#> $levels$residuals_lag42$values
#> [1] NA
#> 
#> $levels$residuals_lag42$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag43
#> $levels$residuals_lag43$values
#> [1] NA
#> 
#> $levels$residuals_lag43$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag44
#> $levels$residuals_lag44$values
#> [1] NA
#> 
#> $levels$residuals_lag44$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag45
#> $levels$residuals_lag45$values
#> [1] NA
#> 
#> $levels$residuals_lag45$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag46
#> $levels$residuals_lag46$values
#> [1] NA
#> 
#> $levels$residuals_lag46$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag47
#> $levels$residuals_lag47$values
#> [1] NA
#> 
#> $levels$residuals_lag47$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag48
#> $levels$residuals_lag48$values
#> [1] NA
#> 
#> $levels$residuals_lag48$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag49
#> $levels$residuals_lag49$values
#> [1] NA
#> 
#> $levels$residuals_lag49$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag50
#> $levels$residuals_lag50$values
#> [1] NA
#> 
#> $levels$residuals_lag50$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag51
#> $levels$residuals_lag51$values
#> [1] NA
#> 
#> $levels$residuals_lag51$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag52
#> $levels$residuals_lag52$values
#> [1] NA
#> 
#> $levels$residuals_lag52$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag53
#> $levels$residuals_lag53$values
#> [1] NA
#> 
#> $levels$residuals_lag53$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag54
#> $levels$residuals_lag54$values
#> [1] NA
#> 
#> $levels$residuals_lag54$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag55
#> $levels$residuals_lag55$values
#> [1] NA
#> 
#> $levels$residuals_lag55$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag56
#> $levels$residuals_lag56$values
#> [1] NA
#> 
#> $levels$residuals_lag56$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag57
#> $levels$residuals_lag57$values
#> [1] NA
#> 
#> $levels$residuals_lag57$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag58
#> $levels$residuals_lag58$values
#> [1] NA
#> 
#> $levels$residuals_lag58$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag59
#> $levels$residuals_lag59$values
#> [1] NA
#> 
#> $levels$residuals_lag59$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag60
#> $levels$residuals_lag60$values
#> [1] NA
#> 
#> $levels$residuals_lag60$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag61
#> $levels$residuals_lag61$values
#> [1] NA
#> 
#> $levels$residuals_lag61$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag62
#> $levels$residuals_lag62$values
#> [1] NA
#> 
#> $levels$residuals_lag62$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag63
#> $levels$residuals_lag63$values
#> [1] NA
#> 
#> $levels$residuals_lag63$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag64
#> $levels$residuals_lag64$values
#> [1] NA
#> 
#> $levels$residuals_lag64$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag65
#> $levels$residuals_lag65$values
#> [1] NA
#> 
#> $levels$residuals_lag65$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag66
#> $levels$residuals_lag66$values
#> [1] NA
#> 
#> $levels$residuals_lag66$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag67
#> $levels$residuals_lag67$values
#> [1] NA
#> 
#> $levels$residuals_lag67$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag68
#> $levels$residuals_lag68$values
#> [1] NA
#> 
#> $levels$residuals_lag68$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag69
#> $levels$residuals_lag69$values
#> [1] NA
#> 
#> $levels$residuals_lag69$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag70
#> $levels$residuals_lag70$values
#> [1] NA
#> 
#> $levels$residuals_lag70$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag71
#> $levels$residuals_lag71$values
#> [1] NA
#> 
#> $levels$residuals_lag71$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag72
#> $levels$residuals_lag72$values
#> [1] NA
#> 
#> $levels$residuals_lag72$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag73
#> $levels$residuals_lag73$values
#> [1] NA
#> 
#> $levels$residuals_lag73$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag74
#> $levels$residuals_lag74$values
#> [1] NA
#> 
#> $levels$residuals_lag74$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag75
#> $levels$residuals_lag75$values
#> [1] NA
#> 
#> $levels$residuals_lag75$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag76
#> $levels$residuals_lag76$values
#> [1] NA
#> 
#> $levels$residuals_lag76$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag77
#> $levels$residuals_lag77$values
#> [1] NA
#> 
#> $levels$residuals_lag77$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag78
#> $levels$residuals_lag78$values
#> [1] NA
#> 
#> $levels$residuals_lag78$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag79
#> $levels$residuals_lag79$values
#> [1] NA
#> 
#> $levels$residuals_lag79$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag80
#> $levels$residuals_lag80$values
#> [1] NA
#> 
#> $levels$residuals_lag80$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag81
#> $levels$residuals_lag81$values
#> [1] NA
#> 
#> $levels$residuals_lag81$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag82
#> $levels$residuals_lag82$values
#> [1] NA
#> 
#> $levels$residuals_lag82$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag83
#> $levels$residuals_lag83$values
#> [1] NA
#> 
#> $levels$residuals_lag83$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag84
#> $levels$residuals_lag84$values
#> [1] NA
#> 
#> $levels$residuals_lag84$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag85
#> $levels$residuals_lag85$values
#> [1] NA
#> 
#> $levels$residuals_lag85$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag86
#> $levels$residuals_lag86$values
#> [1] NA
#> 
#> $levels$residuals_lag86$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag87
#> $levels$residuals_lag87$values
#> [1] NA
#> 
#> $levels$residuals_lag87$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag88
#> $levels$residuals_lag88$values
#> [1] NA
#> 
#> $levels$residuals_lag88$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag89
#> $levels$residuals_lag89$values
#> [1] NA
#> 
#> $levels$residuals_lag89$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals_lag90
#> $levels$residuals_lag90$values
#> [1] NA
#> 
#> $levels$residuals_lag90$ordered
#> [1] NA
#> 
#> 
#> $levels$residuals
#> $levels$residuals$values
#> [1] NA
#> 
#> $levels$residuals$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_month
#> $levels$ds_month$values
#>  [1] "Jan" "Feb" "Mar" "Apr" "May" "Jun" "Jul" "Aug" "Sep" "Oct" "Nov" "Dec"
#> 
#> $levels$ds_month$ordered
#> [1] FALSE
#> 
#> $levels$ds_month$factor
#> [1] TRUE
#> 
#> 
#> $levels$ds_dow
#> $levels$ds_dow$values
#> [1] "Sun" "Mon" "Tue" "Wed" "Thu" "Fri" "Sat"
#> 
#> $levels$ds_dow$ordered
#> [1] FALSE
#> 
#> $levels$ds_dow$factor
#> [1] TRUE
#> 
#> 
#> $levels$ds_quarter
#> $levels$ds_quarter$values
#> [1] NA
#> 
#> $levels$ds_quarter$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_week
#> $levels$ds_week$values
#> [1] NA
#> 
#> $levels$ds_week$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_decimal
#> $levels$ds_decimal$values
#> [1] NA
#> 
#> $levels$ds_decimal$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_sin_1
#> $levels$ds_sin_1$values
#> [1] NA
#> 
#> $levels$ds_sin_1$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_sin_2
#> $levels$ds_sin_2$values
#> [1] NA
#> 
#> $levels$ds_sin_2$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_sin_3
#> $levels$ds_sin_3$values
#> [1] NA
#> 
#> $levels$ds_sin_3$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_sin_4
#> $levels$ds_sin_4$values
#> [1] NA
#> 
#> $levels$ds_sin_4$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_sin_5
#> $levels$ds_sin_5$values
#> [1] NA
#> 
#> $levels$ds_sin_5$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_cos_1
#> $levels$ds_cos_1$values
#> [1] NA
#> 
#> $levels$ds_cos_1$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_cos_2
#> $levels$ds_cos_2$values
#> [1] NA
#> 
#> $levels$ds_cos_2$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_cos_3
#> $levels$ds_cos_3$values
#> [1] NA
#> 
#> $levels$ds_cos_3$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_cos_4
#> $levels$ds_cos_4$values
#> [1] NA
#> 
#> $levels$ds_cos_4$ordered
#> [1] NA
#> 
#> 
#> $levels$ds_cos_5
#> $levels$ds_cos_5$values
#> [1] NA
#> 
#> $levels$ds_cos_5$ordered
#> [1] NA
#> 
#> 
#> 
#> $retained
#> [1] FALSE
#> 
#> $requirements
#> $requirements$bake
#> named logical(0)
#> 
#> 
#> $tr_info
#>   nrows ncomplete
#> 1 92184     47094
#> 
#> $orig_lvls
#> $orig_lvls$ticker
#> $orig_lvls$ticker$values
#>   [1] "A"     "AAL"   "AAP"   "AAPL"  "ABBV"  "ABC"   "ABMD"  "ABT"   "ACN"  
#>  [10] "ADBE"  "ADI"   "ADM"   "ADP"   "ADSK"  "AEE"   "AEP"   "AES"   "AFL"  
#>  [19] "AIG"   "AIZ"   "AJG"   "AKAM"  "ALB"   "ALGN"  "ALK"   "ALL"   "ALLE" 
#>  [28] "AMAT"  "AMCR"  "AMD"   "AME"   "AMGN"  "AMP"   "AMT"   "AMZN"  "ANET" 
#>  [37] "ANSS"  "AON"   "AOS"   "APA"   "APD"   "APH"   "APTV"  "ARE"   "ATO"  
#>  [46] "ATVI"  "AVB"   "AVGO"  "AVY"   "AWK"   "AXP"   "AZO"   "BA"    "BAC"  
#>  [55] "BALL"  "BAX"   "BBWI"  "BBY"   "BDX"   "BIIB"  "BIO"   "BK"    "BKNG" 
#>  [64] "BKR"   "BLK"   "BMY"   "BR"    "BRO"   "BSX"   "BWA"   "BXP"   "CAH"  
#>  [73] "CARR"  "CAT"   "CBOE"  "CBRE"  "CCL"   "CDAY"  "CDNS"  "CDW"   "CE"   
#>  [82] "CHRW"  "CNC"   "CNP"   "COF"   "CPB"   "CPT"   "CTLT"  "CZR"   "GOOG" 
#>  [91] "GOOGL" "KMX"   "LNT"   "MMM"   "MO"    "T"     "TECH"  "WRB"   "C"    
#> [100] "CAG"   "CB"    "CCI"   "CEG"   "CF"    "CFG"   "CHD"   "CHTR"  "CI"   
#> [109] "CINF"  "CL"    "CLX"   "CMA"   "CMCSA" "CME"   "CMG"   "CMI"   "CMS"  
#> [118] "COO"   "COP"   "COST"  "CPRT"  "CRL"   "CSCO"  "CSX"   "CTAS"  "CTRA" 
#> [127] "CTSH"  "CTVA"  "CTXS"  "CVS"   "CVX"   "D"     "DAL"   "DD"    "DE"   
#> [136] "DFS"   "DG"    "DHI"   "DHR"   "DIS"   "DISH"  "DLR"   "DLTR"  "DOV"  
#> [145] "DOW"   "DPZ"   "DRE"   "DRI"   "DTE"   "DUK"   "DVA"   "DVN"   "DXC"  
#> [154] "DXCM"  "EA"    "EBAY"  "ECL"   "ED"    "EFX"   "EIX"   "EL"    "ELV"  
#> [163] "EMN"   "EMR"   "ENPH"  "EOG"   "EPAM"  "EQIX"  "EQR"   "ES"    "ESS"  
#> [172] "ETN"   "ETR"   "ETSY"  "EVRG"  "EW"    "EXC"   "EXPD"  "EXPE"  "EXR"  
#> [181] "FANG"  "FAST"  "FDS"   "FDX"   "FE"    "FFIV"  "FIS"   "FITB"  "FRC"  
#> [190] "FRT"   "GLW"   "KO"    "LLY"   "RE"    "SCHW"  "STZ"   "XOM"   "XRAY" 
#> [199] "BEN"   "F"     "FBHS"  "FCX"   "FISV"  "FLT"   "FMC"   "FOX"   "FOXA" 
#> [208] "FTNT"  "FTV"   "GD"    "GE"    "GILD"  "GIS"   "GL"    "GM"    "GNRC" 
#> [217] "GPC"   "GPN"   "GRMN"  "GS"    "HAL"   "HAS"   "HBAN"  "HCA"   "HD"   
#> [226] "HES"   "HIG"   "HII"   "HLT"   "HOLX"  "HON"   "HPE"   "HPQ"   "HRL"  
#> [235] "HSIC"  "HST"   "HSY"   "HUM"   "HWM"   "IBM"   "ICE"   "IDXX"  "IEX"  
#> [244] "IFF"   "ILMN"  "INCY"  "INTC"  "INTU"  "IP"    "IPG"   "IQV"   "IR"   
#> [253] "IRM"   "ISRG"  "IT"    "ITW"   "IVZ"   "J"     "JBHT"  "JCI"   "JKHY" 
#> [262] "JNJ"   "JNPR"  "JPM"   "K"     "KDP"   "KEY"   "KEYS"  "KHC"   "KIM"  
#> [271] "KLAC"  "KMB"   "KMI"   "KR"    "L"     "LDOS"  "LEN"   "LH"    "LHX"  
#> [280] "LIN"   "LKQ"   "LMT"   "LNC"   "LOW"   "LRCX"  "LUMN"  "LVS"   "LW"   
#> [289] "LYB"   "LYV"   "MAR"   "MKTX"  "MLM"   "MMC"   "MPC"   "MRO"   "MTB"  
#> [298] "PEAK"  "DGX"   "MA"    "MAA"   "MAS"   "MCD"   "MCHP"  "MCK"   "MCO"  
#> [307] "MDLZ"  "MDT"   "MET"   "META"  "MGM"   "MHK"   "MKC"   "MNST"  "MOH"  
#> [316] "MOS"   "MPWR"  "MRK"   "MRNA"  "MS"    "MSCI"  "MSFT"  "MSI"   "MTCH" 
#> [325] "MTD"   "MU"    "NCLH"  "NDAQ"  "NDSN"  "NEE"   "NEM"   "NFLX"  "NI"   
#> [334] "NKE"   "NLOK"  "NLSN"  "NOC"   "NRG"   "NSC"   "NTAP"  "NTRS"  "NUE"  
#> [343] "NVDA"  "NVR"   "NWL"   "NWS"   "NWSA"  "NXPI"  "O"     "ODFL"  "OGN"  
#> [352] "OKE"   "OMC"   "ON"    "ORCL"  "ORLY"  "OTIS"  "OXY"   "PARA"  "PAYC" 
#> [361] "PAYX"  "PCAR"  "PEG"   "PENN"  "PEP"   "PFE"   "PFG"   "PG"    "PGR"  
#> [370] "PH"    "PHM"   "PKG"   "PKI"   "PLD"   "PM"    "PNC"   "PNR"   "PNW"  
#> [379] "POOL"  "PPG"   "PPL"   "PRU"   "PSA"   "PSX"   "PTC"   "PVH"   "PWR"  
#> [388] "PXD"   "PYPL"  "QCOM"  "QRVO"  "REG"   "REGN"  "RF"    "RJF"   "RL"   
#> [397] "RTX"   "TAP"   "CRM"   "GWW"   "LUV"   "NOW"   "RCL"   "RHI"   "RMD"  
#> [406] "ROK"   "ROL"   "ROP"   "ROST"  "RSG"   "SBAC"  "SBNY"  "SBUX"  "SEDG" 
#> [415] "SEE"   "SHW"   "SIVB"  "SJM"   "SLB"   "SNA"   "SNPS"  "SO"    "SPG"  
#> [424] "SPGI"  "SRE"   "STE"   "STT"   "STX"   "SWK"   "SWKS"  "SYF"   "SYK"  
#> [433] "SYY"   "TDG"   "TDY"   "TEL"   "TER"   "TFC"   "TFX"   "TGT"   "TJX"  
#> [442] "TMO"   "TMUS"  "TPR"   "TRMB"  "TROW"  "TRV"   "TSCO"  "TSLA"  "TSN"  
#> [451] "TT"    "TTWO"  "TWTR"  "TXN"   "TXT"   "TYL"   "UAL"   "UDR"   "UHS"  
#> [460] "ULTA"  "UNH"   "UNP"   "UPS"   "URI"   "USB"   "V"     "VFC"   "VICI" 
#> [469] "VLO"   "VMC"   "VNO"   "VRSK"  "VRSN"  "VRTX"  "VTR"   "VTRS"  "VZ"   
#> [478] "WAB"   "WAT"   "WBA"   "WBD"   "WDC"   "WEC"   "WELL"  "WFC"   "WHR"  
#> [487] "WM"    "WMB"   "WMT"   "WRK"   "WST"   "WTW"   "WY"    "WYNN"  "XEL"  
#> [496] "XYL"   "YUM"   "ZBH"   "ZBRA"  "ZION"  "ZTS"  
#> 
#> $orig_lvls$ticker$ordered
#> [1] FALSE
#> 
#> $orig_lvls$ticker$factor
#> [1] TRUE
#> 
#> 
#> $orig_lvls$ds
#> $orig_lvls$ds$values
#> [1] NA
#> 
#> $orig_lvls$ds$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag8
#> $orig_lvls$residuals_lag8$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag8$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag9
#> $orig_lvls$residuals_lag9$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag9$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag10
#> $orig_lvls$residuals_lag10$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag10$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag11
#> $orig_lvls$residuals_lag11$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag11$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag12
#> $orig_lvls$residuals_lag12$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag12$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag13
#> $orig_lvls$residuals_lag13$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag13$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag14
#> $orig_lvls$residuals_lag14$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag14$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag15
#> $orig_lvls$residuals_lag15$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag15$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag16
#> $orig_lvls$residuals_lag16$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag16$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag17
#> $orig_lvls$residuals_lag17$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag17$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag18
#> $orig_lvls$residuals_lag18$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag18$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag19
#> $orig_lvls$residuals_lag19$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag19$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag20
#> $orig_lvls$residuals_lag20$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag20$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag21
#> $orig_lvls$residuals_lag21$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag21$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag22
#> $orig_lvls$residuals_lag22$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag22$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag23
#> $orig_lvls$residuals_lag23$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag23$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag24
#> $orig_lvls$residuals_lag24$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag24$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag25
#> $orig_lvls$residuals_lag25$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag25$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag26
#> $orig_lvls$residuals_lag26$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag26$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag27
#> $orig_lvls$residuals_lag27$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag27$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag28
#> $orig_lvls$residuals_lag28$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag28$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag29
#> $orig_lvls$residuals_lag29$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag29$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag30
#> $orig_lvls$residuals_lag30$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag30$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag31
#> $orig_lvls$residuals_lag31$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag31$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag32
#> $orig_lvls$residuals_lag32$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag32$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag33
#> $orig_lvls$residuals_lag33$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag33$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag34
#> $orig_lvls$residuals_lag34$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag34$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag35
#> $orig_lvls$residuals_lag35$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag35$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag36
#> $orig_lvls$residuals_lag36$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag36$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag37
#> $orig_lvls$residuals_lag37$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag37$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag38
#> $orig_lvls$residuals_lag38$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag38$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag39
#> $orig_lvls$residuals_lag39$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag39$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag40
#> $orig_lvls$residuals_lag40$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag40$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag41
#> $orig_lvls$residuals_lag41$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag41$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag42
#> $orig_lvls$residuals_lag42$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag42$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag43
#> $orig_lvls$residuals_lag43$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag43$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag44
#> $orig_lvls$residuals_lag44$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag44$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag45
#> $orig_lvls$residuals_lag45$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag45$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag46
#> $orig_lvls$residuals_lag46$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag46$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag47
#> $orig_lvls$residuals_lag47$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag47$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag48
#> $orig_lvls$residuals_lag48$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag48$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag49
#> $orig_lvls$residuals_lag49$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag49$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag50
#> $orig_lvls$residuals_lag50$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag50$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag51
#> $orig_lvls$residuals_lag51$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag51$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag52
#> $orig_lvls$residuals_lag52$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag52$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag53
#> $orig_lvls$residuals_lag53$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag53$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag54
#> $orig_lvls$residuals_lag54$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag54$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag55
#> $orig_lvls$residuals_lag55$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag55$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag56
#> $orig_lvls$residuals_lag56$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag56$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag57
#> $orig_lvls$residuals_lag57$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag57$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag58
#> $orig_lvls$residuals_lag58$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag58$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag59
#> $orig_lvls$residuals_lag59$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag59$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag60
#> $orig_lvls$residuals_lag60$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag60$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag61
#> $orig_lvls$residuals_lag61$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag61$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag62
#> $orig_lvls$residuals_lag62$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag62$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag63
#> $orig_lvls$residuals_lag63$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag63$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag64
#> $orig_lvls$residuals_lag64$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag64$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag65
#> $orig_lvls$residuals_lag65$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag65$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag66
#> $orig_lvls$residuals_lag66$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag66$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag67
#> $orig_lvls$residuals_lag67$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag67$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag68
#> $orig_lvls$residuals_lag68$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag68$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag69
#> $orig_lvls$residuals_lag69$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag69$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag70
#> $orig_lvls$residuals_lag70$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag70$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag71
#> $orig_lvls$residuals_lag71$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag71$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag72
#> $orig_lvls$residuals_lag72$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag72$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag73
#> $orig_lvls$residuals_lag73$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag73$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag74
#> $orig_lvls$residuals_lag74$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag74$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag75
#> $orig_lvls$residuals_lag75$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag75$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag76
#> $orig_lvls$residuals_lag76$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag76$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag77
#> $orig_lvls$residuals_lag77$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag77$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag78
#> $orig_lvls$residuals_lag78$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag78$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag79
#> $orig_lvls$residuals_lag79$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag79$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag80
#> $orig_lvls$residuals_lag80$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag80$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag81
#> $orig_lvls$residuals_lag81$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag81$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag82
#> $orig_lvls$residuals_lag82$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag82$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag83
#> $orig_lvls$residuals_lag83$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag83$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag84
#> $orig_lvls$residuals_lag84$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag84$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag85
#> $orig_lvls$residuals_lag85$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag85$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag86
#> $orig_lvls$residuals_lag86$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag86$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag87
#> $orig_lvls$residuals_lag87$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag87$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag88
#> $orig_lvls$residuals_lag88$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag88$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag89
#> $orig_lvls$residuals_lag89$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag89$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals_lag90
#> $orig_lvls$residuals_lag90$values
#> [1] NA
#> 
#> $orig_lvls$residuals_lag90$ordered
#> [1] NA
#> 
#> 
#> $orig_lvls$residuals
#> $orig_lvls$residuals$values
#> [1] NA
#> 
#> $orig_lvls$residuals$ordered
#> [1] NA
#> 
#> 
#> 
#> $last_term_info
#> # A tibble: 101 × 6
#> # Groups:   variable [101]
#>    variable   type    role      source   number skip 
#>    <chr>      <chr>   <list>    <chr>     <dbl> <lgl>
#>  1 ds         date    <chr [1]> original      2 FALSE
#>  2 ds_cos_1   numeric <chr [1]> derived       3 FALSE
#>  3 ds_cos_2   numeric <chr [1]> derived       3 FALSE
#>  4 ds_cos_3   numeric <chr [1]> derived       3 FALSE
#>  5 ds_cos_4   numeric <chr [1]> derived       3 FALSE
#>  6 ds_cos_5   numeric <chr [1]> derived       3 FALSE
#>  7 ds_decimal numeric <chr [1]> derived       3 FALSE
#>  8 ds_dow     nominal <chr [1]> derived       3 FALSE
#>  9 ds_month   nominal <chr [1]> derived       3 FALSE
#> 10 ds_quarter numeric <chr [1]> derived       3 FALSE
#> # … with 91 more rows
#> 
#> attr(,"class")
#> [1] "butchered_recipe" "recipe"          
#> 
#> $extra_role_ptypes
#> NULL
#> 
#> attr(,"class")
#> [1] "default_recipe_blueprint" "recipe_blueprint"        
#> [3] "hardhat_blueprint"       
#> 
#> $mold$extras
#> $mold$extras$roles
#> NULL
#> 
#> 
#> 
#> $case_weights
#> NULL
#> 
#> attr(,"class")
#> [1] "stage_pre" "stage"    
#> 
#> $fit
#> $actions
#> $actions$model
#> $spec
#> $args
#> $args$mtry
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> $args$trees
#> <quosure>
#> expr: ^1000
#> env:  empty
#> 
#> $args$min_n
#> <quosure>
#> expr: ^7
#> env:  empty
#> 
#> $args$tree_depth
#> <quosure>
#> expr: ^12
#> env:  empty
#> 
#> $args$learn_rate
#> <quosure>
#> expr: ^0.0398679046932473
#> env:  empty
#> 
#> $args$loss_reduction
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> $args$sample_size
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> $args$stop_iter
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> 
#> $eng_args
#> <list_of<quosure>>
#> 
#> named list()
#> 
#> $mode
#> [1] "regression"
#> 
#> $user_specified_mode
#> [1] TRUE
#> 
#> $method
#> NULL
#> 
#> $engine
#> [1] "lightgbm"
#> 
#> $user_specified_engine
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "boost_tree" "model_spec"
#> 
#> $formula
#> NULL
#> 
#> attr(,"class")
#> [1] "action_model" "action_fit"   "action"      
#> 
#> 
#> $fit
#> $lvl
#> NULL
#> 
#> $spec
#> $args
#> $args$mtry
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> $args$trees
#> <quosure>
#> expr: ^1000
#> env:  empty
#> 
#> $args$min_n
#> <quosure>
#> expr: ^7
#> env:  empty
#> 
#> $args$tree_depth
#> <quosure>
#> expr: ^12
#> env:  empty
#> 
#> $args$learn_rate
#> <quosure>
#> expr: ^0.0398679046932473
#> env:  empty
#> 
#> $args$loss_reduction
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> $args$sample_size
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> $args$stop_iter
#> <quosure>
#> expr: ^NULL
#> env:  empty
#> 
#> 
#> $eng_args
#> <list_of<quosure>>
#> 
#> named list()
#> 
#> $mode
#> [1] "regression"
#> 
#> $user_specified_mode
#> [1] TRUE
#> 
#> $method
#> $method$libs
#> [1] "lightgbm" "bonsai"  
#> 
#> $method$fit
#> $method$fit$interface
#> [1] "data.frame"
#> 
#> $method$fit$protect
#> [1] "x" "y"
#> 
#> $method$fit$func
#>              pkg              fun 
#>         "bonsai" "train_lightgbm" 
#> 
#> $method$fit$defaults
#> $method$fit$defaults$verbose
#> [1] -1
#> 
#> $method$fit$defaults$num_threads
#> [1] 0
#> 
#> $method$fit$defaults$seed
#> sample.int(10^5, 1)
#> 
#> $method$fit$defaults$deterministic
#> [1] TRUE
#> 
#> 
#> $method$fit$args
#> $method$fit$args$x
#> missing_arg()
#> 
#> $method$fit$args$y
#> missing_arg()
#> 
#> $method$fit$args$num_iterations
#> <quosure>
#> expr: ^1000
#> env:  empty
#> 
#> $method$fit$args$min_data_in_leaf
#> <quosure>
#> expr: ^7
#> env:  empty
#> 
#> $method$fit$args$max_depth
#> <quosure>
#> expr: ^12
#> env:  empty
#> 
#> $method$fit$args$learning_rate
#> <quosure>
#> expr: ^0.0398679046932473
#> env:  empty
#> 
#> $method$fit$args$verbose
#> [1] -1
#> 
#> $method$fit$args$num_threads
#> [1] 0
#> 
#> $method$fit$args$seed
#> sample.int(10^5, 1)
#> 
#> $method$fit$args$deterministic
#> [1] TRUE
#> 
#> 
#> 
#> $method$pred
#> $method$pred$numeric
#> $method$pred$numeric$pre
#> NULL
#> 
#> $method$pred$numeric$post
#> NULL
#> 
#> $method$pred$numeric$func
#>                                   pkg                                   fun 
#>                              "bonsai" "predict_lightgbm_regression_numeric" 
#> 
#> $method$pred$numeric$args
#> $method$pred$numeric$args$object
#> object
#> 
#> $method$pred$numeric$args$new_data
#> new_data
#> 
#> 
#> 
#> 
#> 
#> $engine
#> [1] "lightgbm"
#> 
#> $user_specified_engine
#> [1] TRUE
#> 
#> $defaults
#> $defaults$verbose
#> [1] -1
#> 
#> $defaults$num_threads
#> [1] 0
#> 
#> $defaults$seed
#> sample.int(10^5, 1)
#> 
#> $defaults$deterministic
#> [1] TRUE
#> 
#> 
#> attr(,"class")
#> [1] "boost_tree" "model_spec"
#> 
#> $preproc
#> $preproc$y_var
#> character(0)
#> 
#> 
#> $elapsed
#> $elapsed$elapsed
#> [1] NA
#> 
#> 
#> $fit
#> <lgb.Booster>
#>   Public:
#>     add_valid: function (data, name) 
#>     best_iter: -1
#>     best_score: NA
#>     current_iter: function () 
#>     dump_model: function (num_iteration = NULL, feature_importance_type = 0L) 
#>     eval: function (data, name, feval = NULL) 
#>     eval_train: function (feval = NULL) 
#>     eval_valid: function (feval = NULL) 
#>     finalize: function () 
#>     initialize: function (params = list(), train_set = NULL, modelfile = NULL, 
#>     lower_bound: function () 
#>     params: list
#>     predict: function (data, start_iteration = NULL, num_iteration = NULL, 
#>     raw: NA
#>     record_evals: list
#>     reset_parameter: function (params, ...) 
#>     rollback_one_iter: function () 
#>     save: function () 
#>     save_model: function (filename, num_iteration = NULL, feature_importance_type = 0L) 
#>     save_model_to_string: function (num_iteration = NULL, feature_importance_type = 0L) 
#>     set_train_data_name: function (name) 
#>     to_predictor: function () 
#>     update: function (train_set = NULL, fobj = NULL) 
#>     upper_bound: function () 
#>   Private:
#>     eval_names: NULL
#>     get_eval_info: function () 
#>     handle: lgb.Booster.handle
#>     higher_better_inner_eval: NULL
#>     init_predictor: NULL
#>     inner_eval: function (data_name, data_idx, feval = NULL) 
#>     inner_predict: function (idx) 
#>     is_predicted_cur_iter: list
#>     name_train_set: training
#>     name_valid_sets: list
#>     num_class: 1
#>     num_dataset: 0
#>     predict_buffer: list
#>     set_objective_to_none: FALSE
#>     train_set: NULL
#>     train_set_version: 0
#>     valid_sets: list
#> 
#> attr(,"class")
#> [1] "_lgb.Booster" "model_fit"   
#> 
#> attr(,"class")
#> [1] "stage_fit" "stage"    
#> 
#> $post
#> $actions
#> named list()
#> 
#> attr(,"class")
#> [1] "stage_post" "stage"     
#> 
#> $trained
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "butchered_workflow" "workflow"