Convert a set of files into a data frame

Read one or more files and combine them into a single data frame. If the result is invalid, the default stock data is returned

Usage

input_data(files, default_data = default_stock_data, combine = FALSE)

Arguments

files

A vector of file paths. The function only accepts CSV and Excel files.

default_data

The default data, included for testing purposes. It is not recommended to change this from its default.

combine

If TRUE, the resulting data frame will be combined with the default data frame.

Value

A tibble::tibble().

See also

Examples

file <- tempfile(fileext = "csv")
data <- tibble::tibble(
  x = 1:100
)
readr::write_csv(data, file)

input_data(file)
#> [1] "Files were not converted correctly."
#> # A tibble: 497 × 106
#>    symbol company…¹ excha…² indus…³ website descr…⁴ CEO   secur…⁵ sector prima…⁶
#>    <chr>  <chr>     <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr>    <dbl>
#>  1 MMM    3M Co.    NEW YO… "Offic… www.3m… "3M Co… Mich… 3M Co.  Manag…    3841
#>  2 AOS    A.O. Smi… NEW YO… "Heati… www.ao… "A. O.… Kevi… A.O. S… Manuf…    3630
#>  3 ABT    Abbott L… NEW YO… "Pharm… www.ab… "Abbot… Robe… Abbott… Manuf…    2834
#>  4 ABBV   Abbvie I… NEW YO… "Pharm… www.ab… "AbbVi… Rich… Abbvie… Manuf…    2834
#>  5 ABMD   Abiomed … NASDAQ  "Surgi… www.ab… "Abiom… Mich… Abiome… Manuf…    3841
#>  6 ACN    Accentur… NEW YO… "Data … www.ac… "Accen… NA    Accent… Infor…    7389
#>  7 ATVI   Activisi… NASDAQ  "Softw… www.ac… "Activ… Robe… Activi… Infor…    7372
#>  8 ADM    Archer D… NEW YO… "Flour… www.ad… "ADM u… Juan… Archer… Manuf…    2041
#>  9 ADBE   Adobe Inc NASDAQ  "Softw… www.ad… "Adobe… Shan… Adobe … Infor…    7372
#> 10 ADP    Automati… NASDAQ  "Data … www.ad… "ADP i… Carl… Automa… Infor…    7374
#> # … with 487 more rows, 96 more variables: employees <dbl>, address <chr>,
#> #   address2 <chr>, state <chr>, city <chr>, zip <chr>, country <chr>,
#> #   phone <dbl>, capitalExpenditures <dbl>, cashChange <dbl>, cashFlow <dbl>,
#> #   cashFlowFinancing <dbl>, changesInInventories <dbl>,
#> #   changesInReceivables <dbl>, currency <chr>, depreciation <dbl>,
#> #   filingType <chr>, fiscalDate <dbl>, netBorrowings <dbl>, netIncome <dbl>,
#> #   reportDate <dbl>, totalInvestingCashFlows <dbl>, key <chr>, subkey <chr>, …