Stock price training data

A set of tables containing the stock price and the residuals of the Prophet model over a daily or monthly period. This data was used to train the LightGBM model.

Usage

daily_training_data

monthly_training_data

Format

A tibble with 63,617 / 57,793 rows and 4 variables.

ticker

The ticker/symbol of the stock

ref_date

The date on which the price was recorded

price

The price of the stock (USD)

residuals

The residuals (difference between predictions and actual price) of the Prophet model

An object of class tbl_df (inherits from tbl, data.frame) with 58056 rows and 4 columns.

See also

Examples

daily_training_data
#> # A tibble: 63,252 × 4
#>    ticker ref_date   price_adjusted residuals
#>    <fct>  <date>              <dbl>     <dbl>
#>  1 A      2022-06-13           116.     1.62 
#>  2 A      2022-06-14           116.     1.33 
#>  3 A      2022-06-15           116.     1.03 
#>  4 A      2022-06-16           115.    -0.978
#>  5 A      2022-06-17           112.    -3.26 
#>  6 A      2022-06-21           113.    -2.26 
#>  7 A      2022-06-22           113.    -3.09 
#>  8 A      2022-06-23           118.     1.18 
#>  9 A      2022-06-24           121.     3.82 
#> 10 A      2022-06-27           120.     2.98 
#> # … with 63,242 more rows
monthly_training_data
#> # A tibble: 58,056 × 4
#>    ticker ref_date   price_adjusted residuals
#>    <fct>  <date>              <dbl>     <dbl>
#>  1 A      2012-12-11           26.9     2.26 
#>  2 A      2013-01-02           29.4     3.33 
#>  3 A      2013-02-01           27.2     1.18 
#>  4 A      2013-03-01           27.6     1.14 
#>  5 A      2013-04-01           27.3     1.07 
#>  6 A      2013-05-01           29.9     3.84 
#>  7 A      2013-06-03           28.2     2.21 
#>  8 A      2013-07-01           29.5     0.228
#>  9 A      2013-08-01           30.8     1.05 
#> 10 A      2013-09-03           33.9     4.82 
#> # … with 58,046 more rows